+++ coffeenet-autoconfigure/src/main/java/coffee/synyx/autoconfigure/logging/CoffeeNetLoggingFileProperties.java
@@ -4,6 +4,8 @@ import org.hibernate.validator.constraints.NotBlank;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
+import org.springframework.validation.annotation.Validated;
+
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotNull;
 
@@ -14,6 +16,7 @@ import javax.validation.constraints.NotNull;
  * @author  Tobias Schneider - schneider@synyx.de
  * @since  0.9.0
  */
+@Validated
 @ConfigurationProperties("coffeenet.logging.file")
 public class CoffeeNetLoggingFileProperties {