+++ coffeenet-autoconfigure/src/main/java/coffee/synyx/autoconfigure/security/config/CoffeeNetSecurityResourceProperties.java
@@ -6,6 +6,8 @@ import org.hibernate.validator.constraints.URL;
 import org.springframework.boot.autoconfigure.security.oauth2.resource.ResourceServerProperties;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
+import org.springframework.validation.annotation.Validated;
+
 
 /**
  * This configuration properties class defines all properties that are defined under the prefix
@@ -13,6 +15,7 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
  *
  * @author  Tobias Schneider - schneider@synyx.de
  */
+@Validated
 @ConfigurationProperties("coffeenet.security.resource")
 public class CoffeeNetSecurityResourceProperties extends ResourceServerProperties {