+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ### 0.19.0
 * Upgrade to spring boot 1.5.2
+* Upgrade to spring-cloud-dependencies Camden.SR6
+  * Add @Validated to all @ConfigurationProperties with Validation
 * Change thymeleaf "main" fragment name to "coffeenet-main"
 * Change thymeleaf "scripts" fragment name to "coffeenet-scripts"
 * Prefix thymeleaf intern fragment names with "coffeenet-*"